% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tscourse.R
\name{ARMA.hstep}
\alias{ARMA.hstep}
\title{Computes h-step-ahead predictions from an ARMA(p,q) model}
\usage{
ARMA.hstep(X, h, phi, theta, sigma)
}
\arguments{
\item{X}{a vector containing time series data.}

\item{h}{the number of steps ahead for which to make predictions.}

\item{phi}{a vector with autoregressive coefficients.}

\item{theta}{a vector the moving average coefficients.}

\item{sigma}{the white noise variance.}
}
\value{
a list containing the predicted values as well as the MSPEs of the predictions.
This function builds a matrix of autocovariances for the ARMA(p,q) model using the MA(inf) representation of the process. It then runs the innovations algorithm on this matrix of autocovariances.
}
\description{
Computes h-step-ahead predictions from an ARMA(p,q) model
}
