% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tscourse.R
\name{ARMAacvf}
\alias{ARMAacvf}
\title{Computes the autocovariance function of a causal ARMA(p,q) process.}
\usage{
ARMAacvf(phi, theta, sigma, max.lag = 12, trun = 500)
}
\arguments{
\item{phi}{a vector with autoregressive coefficients.}

\item{theta}{a vector the moving average coefficients.}

\item{sigma}{the white noise variance.}

\item{max.lag}{the number of lags at which to return the value of the autocovariance function.}

\item{trun}{the order at which to truncate the MA(Inf) representation of the causal ARMA(p,q) process.}
}
\value{
A vector containing the values of the autocovariance function at lags \code{0} through \code{max.lag}.
}
\description{
Computes the autocovariance function of a causal ARMA(p,q) process.
}
