% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tscourse.R
\name{ARMAimpute}
\alias{ARMAimpute}
\title{Predicts missing values of a time series based on an ARMA model}
\usage{
ARMAimpute(X, phi, theta, sigma, plot = TRUE)
}
\arguments{
\item{X}{a vector containing time series data with missing values.}

\item{phi}{a vector with autoregressive coefficients.}

\item{theta}{a vector the moving average coefficients.}

\item{sigma}{the white noise variance.}
}
\value{
a list containing the predicted values of the time series, upper and lower bounds for the 95% prediction intervals, and sets of indices corresponding to missing and non-missing observations.
This function predicts or imputes missing values in a time series by constructing an autocovariance function out of given ARMA parameters.  One should find estimated for the ARMA parameters based on the observed data using the \code{arima()} function and then feed these into this function to get the predictions for the missing values.
}
\description{
Predicts missing values of a time series based on an ARMA model
}
