% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tscourse.R
\name{ARMAtoMAinf}
\alias{ARMAtoMAinf}
\title{Gives coefficients of causal representation of a causal ARMA(p,q) model}
\usage{
ARMAtoMAinf(phi, theta, trun)
}
\arguments{
\item{phi}{a vector with autoregressive coefficients.}

\item{theta}{a vector the moving average coefficients.}

\item{trun}{the number of terms to keep in the causal representation of the model}
}
\value{
a vector containing the first \code{trun+1} moving average coefficients in the causal representation of the ARMA(p,q) model
This function recursively computes the coefficients in the MA(Inf) representation of the causal ARMA(p,q) model.
}
\description{
Gives coefficients of causal representation of a causal ARMA(p,q) model
}
