% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tscourse.R
\name{ARMAtoSD}
\alias{ARMAtoSD}
\title{Find the spectral density function of an ARMA(p,q) process.}
\usage{
ARMAtoSD(phi = NULL, theta = NULL, sigma, plot = TRUE)
}
\arguments{
\item{phi}{a vector with autoregressive coefficients.}

\item{theta}{a vector the moving average coefficients.}

\item{sigma}{the white noise variance.}

\item{plot}{if \code{TRUE} the spectral density is plotted}
}
\value{
a list containing values of the spectral density function and the corresponding frequencies
}
\description{
Find the spectral density function of an ARMA(p,q) process.
}
