% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tscourse.R
\name{get.ARMA.data}
\alias{get.ARMA.data}
\title{Generates data from an ARMA(p,q) model with iid Normal innovations.}
\usage{
get.ARMA.data(phi = NULL, theta = NULL, sigma = 1, n, trun = 500)
}
\arguments{
\item{phi}{a vector with autoregressive coefficients.}

\item{theta}{a vector the moving average coefficients.}

\item{sigma}{the white noise variance.}

\item{n}{the length of the realization to be generated.}

\item{trun}{the number of terms to keep in the causal representation of the model, which is used to generate the data.}
}
\value{
a length-\code{n} realization of the time series.

This function generates a length-\code{n} realization from a causal invertible ARMA(p,q) model with iid Normal innovations.
}
\description{
Generates data from an ARMA(p,q) model with iid Normal innovations.
}
