% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tscourse.R
\name{parzen}
\alias{parzen}
\title{Evaluate the Parzen window}
\usage{
parzen(x)
}
\arguments{
\item{x}{a real number.}
}
\value{
the value of the Parzen window function.
}
\description{
Evaluate the Parzen window
}
