% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tscourse.R
\name{pgram}
\alias{pgram}
\title{Compute the periodogram}
\usage{
pgram(X, plot = FALSE)
}
\arguments{
\item{X}{a vector of data}

\item{plot}{if TRUE a plot of the periodogram is generated}
}
\value{
a list containing the periodogram ordinates, the Fourier frequencies, and the cumulative periodogram

This function computes the periodogram at the Fourier frequencies.
}
\description{
Compute the periodogram
}
