% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tscourse.R
\name{sample.acf}
\alias{sample.acf}
\title{Computes the sample autocovariance and autocorrelation functions.}
\usage{
sample.acf(x, max.lag = 12)
}
\arguments{
\item{x}{a vector containing time series data.}

\item{max.lag}{the largest lag at which to compute the autocovariance and autocorrelation functions.}
}
\value{
a list containing the autocovariance at lags zero through \code{max.lag}
This function computes the sample autocovariance and autocorrelation functions based on the input data.
}
\description{
Computes the sample autocovariance and autocorrelation functions.
}
