% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tscourse.R
\name{SDlagWest}
\alias{SDlagWest}
\title{Compute a lag-windor estimator of the spectral density}
\usage{
SDlagWest(X, L, window = parzen, nlambda = 10000, plot = TRUE)
}
\arguments{
\item{X}{a numeric vector containing the observed data}

\item{L}{the number of lags at which to truncate the autocovariance function.}

\item{window}{the lag window function to be used. Default is the Parzen window.}

\item{nlambda}{the number of values between \code{-pi} and \code{pi} for which the estimate of the spectral density should be computed.}

\item{plot}{if \code{TRUE} then a plot of the estimated spectral density is produced.}
}
\value{
a list containing values of the estimated spectral density function and the corresponding frequencies.
}
\description{
Compute a lag-windor estimator of the spectral density
}
