% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tscourse.R
\name{SDtoMAinf}
\alias{SDtoMAinf}
\title{Find the moving average representation of a time series based on the spectral density.}
\usage{
SDtoMAinf(f, lambda, trun = 500, tol = 1e-04)
}
\arguments{
\item{f}{a vector with evaluations of the spectral density}

\item{lambda}{the frequencies to which the values of f correspond. Should be evenly spaced and dense between \code{-pi} and \code{pi}.}

\item{trun}{the number of terms to keep in the moving average representation of the time series}

\item{tol}{controls the accuracy. Smaller values require more computation time.}
}
\value{
a vector containing the coefficients of the moving average representation of the time series.
This is based on the work of Pourahmadi (1984).
}
\description{
Find the moving average representation of a time series based on the spectral density.
}
\references{
Pourahmadi, M. (1984). Taylor expansion of and some applications. \emph{The American Mathematical Monthly}, 91(5), 303-307.
}
