\name{tscourse-package}
\alias{tscourse-package}
\alias{tscourse}
\docType{package}
\title{
\packageTitle{tscourse}
}
\description{
\packageDescription{tscourse}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{tscourse}
\packageIndices{tscourse}
This package provides basic functions for a graduate-level time series course I taught during the Spring semester of 2019.
}
\author{
\packageAuthor{tscourse}

Maintainer: \packageMaintainer{tscourse}
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
