% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tscourse.R
\name{unitroottest.DF}
\alias{unitroottest.DF}
\title{Perform the Dickey-Fuller unit-root test}
\usage{
unitroottest.DF(X, p = 1)
}
\arguments{
\item{X}{a vector containing time series data.}

\item{p}{the autoregressive order on which to base the test.}
}
\value{
a list with logicals indicating rejection or failure to reject at the 0.05 and 0.01 significance levels as well as the value of the Dickey-Fuller test statistic.

This function performs the Dickey-Fuller unit-root test.
}
\description{
Perform the Dickey-Fuller unit-root test
}
