% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tscourse.R
\name{WNtest.Bartlett}
\alias{WNtest.Bartlett}
\title{Perform Bartlett's test for whether a time series is iid Normal.}
\usage{
WNtest.Bartlett(X, plot = FALSE)
}
\arguments{
\item{X}{a vector containing time series data.}

\item{plot}{if TRUE a plot of the cumulative periodogram is generated, on which the Kolmogorov-Smirnov bounds are displayed.}
}
\value{
the p value of Bartlett's test

This function performs Bartlett's test for whether the time series is iid Normal.
}
\description{
Perform Bartlett's test for whether a time series is iid Normal.
}
