% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tscourse.R
\name{WNtest.LB}
\alias{WNtest.LB}
\title{Perform the Ljung-Box test for whether a time series is iid.}
\usage{
WNtest.LB(X, k = 1, nparms = 0)
}
\arguments{
\item{X}{a vector containing time series data.}

\item{k}{number of lags on which to base the test statistic}

\item{nparms}{number of parameters in the fitted model; the degrees of freedom of the chi-squared distribution from which the p value is computed will be set equal to \code{k - nparms}.}
}
\value{
the p value of the Ljung-Box test

This function performs the Ljung-Box test for whether the time series is iid.
}
\description{
Perform the Ljung-Box test for whether a time series is iid.
}
