% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tscourse.R
\name{DL.1step}
\alias{DL.1step}
\title{Performs the Durbin-Levinson algorithm for one-step-ahead prediction.}
\usage{
DL.1step(X, gamma.0, gamma.n)
}
\arguments{
\item{X}{a vector containing time series data.}

\item{gamma.0}{the value of the autocovariance function at lag zero}

\item{gamma.n}{a vector containing the values of the autocovariance function at lags \code{1} through \code{length(X)}}
}
\value{
a list containing the one-step-ahead predictions as well as the values of the partial autocorrelation function at lags \code{1} through \code{length(X)}
This function performs the Durbin-Levinson algorithm for one-step-ahead prediction
}
\description{
Performs the Durbin-Levinson algorithm for one-step-ahead prediction.
}
