% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tscourse.R
\name{innov.hstep}
\alias{innov.hstep}
\title{Performs the innovations algorithm for h-step-ahead prediction.}
\usage{
innov.hstep(X, h, K)
}
\arguments{
\item{X}{a vector containing time series data.}

\item{h}{the number of steps ahead for which to make predictions.}

\item{K}{the covariance matrix of the random variables X_1,\dots,X_{n+h}.}
}
\value{
a list containing the predicted values as well as the MSPEs of the predictions and the matrix \code{Theta}.
This function performs the innovations algorithm for one-step-ahead and h-step-ahead prediction. Data are centered before the prediction are computed and then mean is added back to the predictions.
}
\description{
Performs the innovations algorithm for h-step-ahead prediction.
}
